import campagne e multi squad

+++ upload.php
@@ -2,10 +2,7 @@
 ini_set("max_execution_time", "1000");
 set_time_limit(1000);
 
-// default extension accepted
-$extensions= array("xlsx","txt","xls","xlsm", "doc", "docx");
-
-if(isset($_FILES['file'])){
+   if(isset($_FILES['file'])){
       
       $id_upload = $_GET['id_upload'];
       if(isset($_GET['comunicazione'])){
@@ -14,10 +11,6 @@ if(isset($_FILES['file'])){
       }
       elseif(isset($_GET['canale'])){
          $dir = "file/$id_upload/canale/";
-      }
-      elseif(isset($_GET['import_campagne'])){
-         $dir = "file/$id_upload/import_campagne/";
-         $extensions= array("xlsx");
       }
       
       mkdir($dir, 0777, TRUE);
@@ -40,7 +33,7 @@ if(isset($_FILES['file'])){
       
       $file_name = preg_replace("/[^a-z0-9\_\-\.]/i", '_', $file_basename);
 
-      
+      $extensions= array("xlsx","txt","xls","xlsm", "doc", "docx");
       
       if(in_array($file_ext,$extensions)=== false){
          $errors[]="extension not allowed!!!";
@@ -69,7 +62,10 @@ if(isset($_FILES['file'])){
       $id_upload = $_POST['id_upload'];
       $file_name = $id_upload.'/'.$subdir.'/'. $file_basename;
       echo 'deleteeee  '.$file_name;
-      unlink("file/".$file_name); 
+      if (file_exists("file/".$file_name)){
+         unlink("file/".$file_name); 
+      }
+      
   }
   //download file
   elseif(isset($_GET['download'])){
@@ -82,28 +78,25 @@ if(isset($_FILES['file'])){
          elseif(isset($_GET['canale'])){
             $dir = "file\\$id_upload\\canale\\";   
          }
-         elseif(isset($_GET['import_campagne'])){
-            $dir = "file\\$id_upload\\import_campagne\\";   
-         }
          else{
             exit('Download Error !!!');
          }
-         $file_path = $dir.$filename;
-         // verifico che il file esista
-         if (!file_exists($file_path))
-         {
-         // se non esiste stampo un errore
-         echo "Il file non esiste!";
-         }else{
-         // Se il file esiste...
-         // Imposto gli header della pagina per forzare il download del file
-         header("Cache-Control: public");
-         header("Content-Description: File Transfer");
-         header("Content-Disposition: attachment; filename= " . $filename);
-         header("Content-Transfer-Encoding: binary");
-         // Leggo il contenuto del file
-         readfile($file_path);
-         }
+$file_path = $dir.$filename;
+  // verifico che il file esista
+if (!file_exists($file_path))
+{
+  // se non esiste stampo un errore
+  echo "Il file non esiste!";
+}else{
+  // Se il file esiste...
+  // Imposto gli header della pagina per forzare il download del file
+  header("Cache-Control: public");
+  header("Content-Description: File Transfer");
+  header("Content-Disposition: attachment; filename= " . $filename);
+  header("Content-Transfer-Encoding: binary");
+  // Leggo il contenuto del file
+  readfile($file_path);
+}
   
       
 }
